% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundoc.R
\name{fundoc_with_ud_macro}
\alias{fundoc_with_ud_macro}
\title{Function With User Defined Macro In Description}
\usage{
fundoc_with_ud_macro(Rd)
}
\arguments{
\item{Rd}{Object of class \code{Rd}}
}
\value{
Some return value
}
\description{
CRANpkg macro: \CRANpkg{mycranpkg}
}
