% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundoc.R
\name{Person}
\alias{Person}
\title{This is my Person class}
\description{
Person class description
}
\examples{
Person$new(name="Bill", hair="Blond")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the person}

\item{\code{hair}}{Hair colour}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Person-new}{\code{Person$new()}}
\item \href{#method-Person-set_hair}{\code{Person$set_hair()}}
\item \href{#method-Person-clone}{\code{Person$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-new"></a>}}
\if{latex}{\out{\hypertarget{method-Person-new}{}}}
\subsection{Method \code{new()}}{
Create a person
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$new(name = NA, hair = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the person}

\item{\code{hair}}{Hair colour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-set_hair"></a>}}
\if{latex}{\out{\hypertarget{method-Person-set_hair}{}}}
\subsection{Method \code{set_hair()}}{
Set hair
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$set_hair(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Hair colour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Person-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
